refactor new config property

+++ core/src/main/java/io/questdb/cairo/CairoConfiguration.java
@@ -724,5 +724,5 @@ public interface CairoConfiguration {
 
     boolean useFastAsOfJoin();
 
-    int getWriteBackOffTimeoutOnMemPressureMs();
+    long getWriteBackOffTimeoutOnMemPressureMs();
 }
